requirements.txt
bidict==0.21.3
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-MySQLdb==0.2.0
Flask-SocketIO==5.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
python-dotenv==0.19.0
python-engineio==4.2.1
python-socketio==5.4.0
requests==2.26.0
urllib3==1.26.6
Werkzeug==2.0.1
